compile launcher on windows
moves all the incompatible syscalls behind build tags, but leaves them as TODOs.
