+++ pkg/debug/signal_debug_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package debug
+
+import "github.com/go-kit/kit/log"
+
+func AttachDebugHandler(addrPath string, logger log.Logger) {
+	// TODO: noop for now
+}
