+++ pkg/osquery/runtime/processgroup_windows.go
@@ -1,0 +1,17 @@
+// +build windows
+
+package runtime
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func setpgid() *syscall.SysProcAttr {
+	return &syscall.SysProcAttr{}
+}
+
+func killProcessGroup(cmd *exec.Cmd) error {
+	// TODO: implement
+	return nil
+}
